# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WhatsAppMessageInput {
  /** @minLength 1 */
  phoneNumberId: string;
  /** @minLength 7 */
  to: string;
  /** @minLength 1 */
  message: string;
}

export interface WhatsAppMessageResult {
  accepted: boolean;
  /** @nullable */
  messageId: string | null;
  recipient: string;
  /** @nullable */
  detail: string | null;
}

export type WhatsAppStatusStatus = typeof WhatsAppStatusStatus[keyof typeof WhatsAppStatusStatus];


export const WhatsAppStatusStatus = {
  accepted: 'accepted',
  sent: 'sent',
  delivered: 'delivered',
  read: 'read',
  failed: 'failed',
} as const;

export interface WhatsAppStatus {
  messageId: string;
  status: WhatsAppStatusStatus;
  /** @nullable */
  detail: string | null;
  updatedAt: string;
}

export interface WhatsAppStatusResponse {
  statuses: WhatsAppStatus[];
}

export interface ErrorResponse {
  error: string;
}

export type GetWhatsAppMessageStatusesParams = {
/**
 * Comma-separated WhatsApp message IDs
 */
messageIds: string;
};

export type VerifyWhatsAppWebhookParams = {
'hub.mode'?: string;
'hub.verify_token'?: string;
'hub.challenge'?: string;
};