# lib/api-zod/src/generated/types/whatsAppMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppMessageInput {
  /** @minLength 1 */
  phoneNumberId: string;
  /** @minLength 7 */
  to: string;
  /** @minLength 1 */
  message: string;
}