# lib/api-zod/src/generated/types/whatsAppMessageResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppMessageResult {
  accepted: boolean;
  /** @nullable */
  messageId: string | null;
  recipient: string;
  /** @nullable */
  detail: string | null;
}