# lib/api-zod/src/generated/types/whatsAppStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppStatusStatus } from './whatsAppStatusStatus';

export interface WhatsAppStatus {
  messageId: string;
  status: WhatsAppStatusStatus;
  /** @nullable */
  detail: string | null;
  updatedAt: Date;
}